backend/my-new-project/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mental-health-monitor"
version = "0.1.0"
description = "LangChain + LangGraph multi-agent mental health behavioral monitoring system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.1",
    "langchain-groq>=0.1.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "openpyxl>=3.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
